@@ -14,13 +14,17 @@
     "@testing-library/user-event": "^12.8.3",
     "bootstrap": "^5.1.3",
     "font-awesome": "^4.7.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.4.0",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "4.0.3",
-    "sass": "^1.52.1",
-    "web-vitals": "^1.1.2"
+    "sass": "^1.69.0",
+    "web-vitals": "^3.5.0"
+  },
+  "devDependencies": {
+    "parcel": "^2.9.3",
+    "process": "^0.11.10"
   },
   "scripts": {
     "start": "react-scripts start",
